Update code to be more OOP and self-contained
